requirements.txt
flask-mail
python-dotenv